+++ docs/changelog.rst
@@ -18,6 +18,8 @@ Version 0.5
 - Deprecate built-in cache. Read more on `issue#23`_.
 - **Redesigned OAuth 1 Flask server**. Read the docs :ref:`flask_oauth1_server`.
 - Deprecate ``client_model``. Read more on `issue#27`_.
+- **Breaking change** on ``AuthorizationCodeGrant.create_authorization_code``,
+  last parameter is changed to an `OAuth2Request` instance.
 
 .. _`issue#23`: https://github.com/lepture/authlib/issues/23
 .. _`issue#27`: https://github.com/lepture/authlib/issues/27